src/main/java/me/toy/netty/telnet/handler/TelnetServerHandler.java
package me.toy.netty.telnet.handler;

import io.netty.channel.*;

@ChannelHandler.Sharable
public class TelnetServerHandler extends SimpleChannelInboundHandler<String> {

    @Override
    public void channelActive(ChannelHandlerContext ctx) throws Exception {
        ctx.write(ResponseGenerator.hello());
        ctx.flush();
    }

    @Override
    protected void channelRead0(ChannelHandlerContext ctx, String request) throws Exception {
        ResponseGenerator generator = new ResponseGenerator(request);

        String response = generator.response();

        ChannelFuture future = ctx.write(response);

        if (generator.isClose()) {
            future.addListener(ChannelFutureListener.CLOSE);
        }
    }

    @Override
    public void channelReadComplete(ChannelHandlerContext ctx) throws Exception {
        ctx.flush();
    }

    @Override
    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) throws Exception {
        cause.printStackTrace();
        ctx.flush();
    }
}
